services/qwen-image/requirements.txt
# Core dependencies
torch>=2.2.0
diffusers>=0.30.0

# HuggingFace
huggingface-hub>=0.19.0

# Web framework
fastapi>=0.109.1
uvicorn>=0.23.0
pydantic>=2.4.0

# Image processing
Pillow>=10.0.0

# System utilities
psutil>=5.9.0
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
